[#482] reputation/router: Move router to `./common`
