+++ pkg/services/reputation/common/router/calls.go
@@ -1,4 +1,4 @@
-package reputationroute
+package router
 
 import (
 	"sync"
