+++ pkg/services/reputation/common/router/opts.go
@@ -1,4 +1,4 @@
-package reputationroute
+package router
 
 import (
 	"github.com/nspcc-dev/neofs-node/pkg/util/logger"
