+++ pkg/services/reputation/common/router/util.go
@@ -1,4 +1,4 @@
-package reputationroute
+package router
 
 import (
 	"bytes"
