+++ pkg/services/reputation/local/managers/deps.go
@@ -2,7 +2,7 @@ package managers
 
 import (
 	"github.com/nspcc-dev/neofs-node/pkg/services/reputation"
-	reputationroute "github.com/nspcc-dev/neofs-node/pkg/services/reputation/local/route"
+	reputationrouter "github.com/nspcc-dev/neofs-node/pkg/services/reputation/common/router"
 )
 
 // ManagerBuilder defines an interface for providing a list
@@ -10,5 +10,5 @@ import (
 type ManagerBuilder interface {
 	// BuildManagers must compose list of managers. It depends on
 	// particular epoch and PeerID of the current route point.
-	BuildManagers(epoch uint64, p reputation.PeerID) ([]reputationroute.ServerInfo, error)
+	BuildManagers(epoch uint64, p reputation.PeerID) ([]reputationrouter.ServerInfo, error)
 }
